+++ src/app/header/header.component.html
@@ -24,7 +24,7 @@
   </h2>
   <div class="header_search">
     <mat-icon> place</mat-icon>
-    <app-city-filter (selectedCity)="onSelectedCity($event)" class="header_city"></app-city-filter>
+    <app-city-filter class="header_city"></app-city-filter>
     <app-searchbar class="header_bar"></app-searchbar>
   </div>
 </div>
