+++ src/app/shared/components/city-autocomplete/city-autocomplete.component.html
@@ -1,0 +1,14 @@
+<form class="search-city">
+  <mat-form-field appearance="none">
+    <input type="text"
+      matInput
+      [formControl]="cityControl"
+      [matAutocomplete]="auto"
+      >
+    <mat-autocomplete autoActiveFirstOption #auto="matAutocomplete" (optionSelected)="onSelect($event)">
+      <mat-option [disabled]= "noCity" *ngFor="let city of filteredCities$ | async" [value]="city">
+          {{city}}
+      </mat-option>
+    </mat-autocomplete>
+  </mat-form-field>
+</form>
