+++ src/app/shared/components/city-autocomplete/city-autocomplete.component.scss
@@ -1,0 +1,13 @@
+:host ::ng-deep .mat-form-field-wrapper{
+    padding: 0;
+   }
+  :host ::ng-deep .mat-form-field-infix{
+    border: none;
+   }
+  .search-city{
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    height: 100%;
+  }
+
