+++ src/app/shared/components/filters-list/city-filter/city-filter.component.html
@@ -1,14 +1,1 @@
-<form class="search-city">
-  <mat-form-field appearance="none">
-    <input type="text"
-      matInput
-      [formControl]="cityControl"
-      [matAutocomplete]="auto"
-      >
-    <mat-autocomplete autoActiveFirstOption #auto="matAutocomplete" (optionSelected)="onSelect($event)">
-      <mat-option [disabled]= "noCity" *ngFor="let city of filteredCities$ | async" [value]="city">
-          {{city}}
-      </mat-option>
-    </mat-autocomplete>
-  </mat-form-field>
-</form>
+<app-city-autocomplete (selectedCity)="onSelectedCity($event)"></app-city-autocomplete>
