--- src/app/shared/components/filters-list/city-filter/city-filter.component.scss
@@ -1,12 +1,0 @@
-:host ::ng-deep .mat-form-field-wrapper{
-  padding: 0;
- }
-:host ::ng-deep .mat-form-field-infix{
-  border: none;
- }
-.search-city{
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  height: 100%;
-}
